+++ skforecast/__init__.py
@@ -1,2 +1,2 @@
 name = "skforecast"
-__version__ = "0.4.2dev"
+__version__ = "0.4.2"