+++ docs/releases/releases.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 
 + Function `backtesting_forecaster()` do not modify the original forecaster.
 
-+ Removed argument `set_out_sample_residuals` in function `backtesting_forecaster()`.
++ Deprecated argument `set_out_sample_residuals` in function `backtesting_forecaster()`.
 
 + Function `model_selection.time_series_spliter` renamed to `model_selection.time_series_splitter`